Create test project and move the tests to it
Create TagsCloudVisualizationTests.csproj
