+++ cs/tdd.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples", "Samples\Samples.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{DF9FE88F-225B-4CE4-A6BE-6F3092867E15}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualizationTests", "TagsCloudVisualizationTests\TagsCloudVisualizationTests.csproj", "{9C6AF192-63A0-4D8F-9195-09B59AE02A7E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{DF9FE88F-225B-4CE4-A6BE-6F3092867E15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DF9FE88F-225B-4CE4-A6BE-6F3092867E15}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DF9FE88F-225B-4CE4-A6BE-6F3092867E15}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C6AF192-63A0-4D8F-9195-09B59AE02A7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C6AF192-63A0-4D8F-9195-09B59AE02A7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C6AF192-63A0-4D8F-9195-09B59AE02A7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C6AF192-63A0-4D8F-9195-09B59AE02A7E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
